updated ingest cases, auto base station and sensor module db creation

+++ ingest/ingest/models/sensor.py
@@ -1,10 +1,10 @@
 from ingest.models.base import Base
-from sqlalchemy import Column, Integer, String, Float, DateTime
+from sqlalchemy import BigInteger, Column, Integer, String, Float, DateTime
 
 class Sensor(Base):
     __tablename__ = 'sensor'
     
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     bsid = Column(Integer, nullable=False)
     smid = Column(Integer, nullable=False)
     name = Column(String, nullable=False)
